pickup fixes
- added audio clip to each pickup
- fixed trigger on cabbage pickup
- added glow and explosion to cabbage and tomato (though tomato glow isn't showing)
