+++ src/Assets/Prefabs/CarrotPickup.prefab
@@ -10009,4 +10009,5 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 7463f603385b4b44ca47184068f5f3d4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: Assembly-CSharp::PowerUp
-  boostMultiplier: 1.5
+  staminaAmount: 15
+  pickupSfx: {fileID: 8300000, guid: 51d24086c38dce14f8cbef9c1dac1fad, type: 3}
